+++ source/victim/square_multiply/square_multiply.h
@@ -3,6 +3,6 @@
 
 int number_of_bits(unsigned int n);
 
-unsigned int square_multiply(unsigned int b, unsigned int e, unsigned int m, int num_bits);
+unsigned long long square_multiply(unsigned int b, unsigned int e, unsigned int m, int num_bits);
 
 #endif
